Save heavy responses to a file for fetch_item_aspects
