+++ lib/ebay_api/middlewares.rb
@@ -1,4 +1,4 @@
 require_relative "middlewares/json_response"
-require_relative "middlewares/gzipped_response"
+require_relative "middlewares/save_to_file_response"
 require_relative "middlewares/paginated_collection"
 require_relative "middlewares/log_request"
