+++ lib/ebay_api/operations/commerce/taxonomy/category_tree/fetch_item_aspects.rb
@@ -4,7 +4,7 @@ class EbayAPI
       scope :category_tree do
         # @see https://developer.ebay.com/api-docs/commerce/taxonomy/resources/category_tree/methods/fetchItemAspects
         operation :fetch_item_aspects do
-          middleware { [GzippedResponse] }
+          middleware { SaveToFileResponse } # returns { "filename": "..." }
 
           path { "fetch_item_aspects" }
           http_method :get
